refactor(http/jwt): aligned structure of common.go with http package
- header names and titles should be moved into http/common to reuse here
- common.go of this package should then be deleted
