+++ internal/app/http/jwt/common.go
@@ -1,17 +1,5 @@
 package jwt
 
-import (
-	"encoding/json"
-	"fmt"
-	"log/slog"
-	"net/http"
-
-	"github.com/web-of-things-open-source/tm-catalog-cli/internal/app/http/server"
-)
-
-// TODO(pedram): refactor common.go into its own package to be reused here
-const HTTPHeaderAuthorization = "Authorization"
-
 // TODO(pedram): refactor common.go into its own package to be reused here
 const (
 	error400Title  = "Bad Request"
@@ -22,6 +10,7 @@ const (
 	error500Title  = "Internal Server Error"
 	error500Detail = "An unhandled error has occurred. Try again later. If it is a bug we already recorded it. Retrying will most likely not help"
 
+	headerAuthorization       = "Authorization"
 	headerContentType         = "Content-Type"
 	headerCacheControl        = "Cache-Control"
 	headerXContentTypeOptions = "X-Content-Type-Options"
@@ -37,19 +26,3 @@ const (
 	ctxUrlRoot      = "urlContextRoot"
 	ctxRelPathDepth = "relPathDepth"
 )
-
-// TODO(pedram): refactor common.go into its own package to be reused here
-func writeErrorResponse(w http.ResponseWriter, err error, status int) {
-	slog.Default().Debug("jwt: " + err.Error())
-	w.Header().Set(headerContentType, mimeJSON)
-	w.WriteHeader(status)
-
-	errString := fmt.Sprint(err)
-	errorResponse := server.ErrorResponse{
-		Status: status,
-		Detail: &errString,
-		Title:  error401Title,
-	}
-	body, _ := json.MarshalIndent(errorResponse, "", "    ")
-	_, _ = w.Write(body)
-}
